reject nil/empty keys and nil values (#99)
Fixes #2.

+++ cleveldb.go
@@ -52,7 +52,9 @@ func NewCLevelDB(name string, dir string) (*CLevelDB, error) {
 
 // Get implements DB.
 func (db *CLevelDB) Get(key []byte) ([]byte, error) {
-	key = nonNilBytes(key)
+	if len(key) == 0 {
+		return nil, errKeyEmpty
+	}
 	res, err := db.db.Get(db.ro, key)
 	if err != nil {
 		return nil, err
@@ -71,8 +73,12 @@ func (db *CLevelDB) Has(key []byte) (bool, error) {
 
 // Set implements DB.
 func (db *CLevelDB) Set(key []byte, value []byte) error {
-	key = nonNilBytes(key)
-	value = nonNilBytes(value)
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
+	if value == nil {
+		return errValueNil
+	}
 	if err := db.db.Put(db.wo, key, value); err != nil {
 		return err
 	}
@@ -81,8 +87,12 @@ func (db *CLevelDB) Set(key []byte, value []byte) error {
 
 // SetSync implements DB.
 func (db *CLevelDB) SetSync(key []byte, value []byte) error {
-	key = nonNilBytes(key)
-	value = nonNilBytes(value)
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
+	if value == nil {
+		return errValueNil
+	}
 	if err := db.db.Put(db.woSync, key, value); err != nil {
 		return err
 	}
@@ -91,7 +101,9 @@ func (db *CLevelDB) SetSync(key []byte, value []byte) error {
 
 // Delete implements DB.
 func (db *CLevelDB) Delete(key []byte) error {
-	key = nonNilBytes(key)
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
 	if err := db.db.Delete(db.wo, key); err != nil {
 		return err
 	}
@@ -100,7 +112,9 @@ func (db *CLevelDB) Delete(key []byte) error {
 
 // DeleteSync implements DB.
 func (db *CLevelDB) DeleteSync(key []byte) error {
-	key = nonNilBytes(key)
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
 	if err := db.db.Delete(db.woSync, key); err != nil {
 		return err
 	}
@@ -164,12 +178,18 @@ func (db *CLevelDB) NewBatch() Batch {
 
 // Iterator implements DB.
 func (db *CLevelDB) Iterator(start, end []byte) (Iterator, error) {
+	if (start != nil && len(start) == 0) || (end != nil && len(end) == 0) {
+		return nil, errKeyEmpty
+	}
 	itr := db.db.NewIterator(db.ro)
 	return newCLevelDBIterator(itr, start, end, false), nil
 }
 
 // ReverseIterator implements DB.
 func (db *CLevelDB) ReverseIterator(start, end []byte) (Iterator, error) {
+	if (start != nil && len(start) == 0) || (end != nil && len(end) == 0) {
+		return nil, errKeyEmpty
+	}
 	itr := db.db.NewIterator(db.ro)
 	return newCLevelDBIterator(itr, start, end, true), nil
 }
